feat: MAX-Квиз bot с HTTP-клиентом и исправлениями
- Добавлен http_client.py для прямых запросов к MAX API
- Реализован KeyboardAdapter с fallback (HTTP → maxapi)
- Исправлены endpoint'ы (без /v1/ префикса)
- Исправлена авторизация (токен без Bearer)
- Исправлен Game.id (Integer + autoincrement для SQLite)
- Добавлены скрипты для загрузки вопросов (load_questions.py)
- Добавлены скрипты для распределения категорий (distribute_questions.py)
- Добавлены скрипты для исправления wrong_answers (fix_wrong_answers.py)
- Добавлена документация (docs/)
- Обновлены модели (models.py) и БД (db.py)

Fixes:
- 404 Path /v1/messages not recognized
- 400 Empty request body
- 401 No access token
- IntegrityError games.id NOT NULL
- ChatButton chat_title Field required

⚠️ Датасеты исключены из git (см. .gitignore)

+++ models.py
@@ -75,8 +75,8 @@ class Question(Base):
     """Модель вопроса."""
     __tablename__ = "questions"
 
-    # ИСПРАВЛЕНО: autoincrement=True для SQLite совместимости
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    # ✅ ИСПРАВЛЕНО: Integer вместо BigInteger для SQLite autoincrement
+    id = Column(Integer, primary_key=True, autoincrement=True)
     text = Column(Text, nullable=False)
     category = Column(Enum(QuestionCategory), nullable=False)
     difficulty = Column(Enum(DifficultyLevel), nullable=False)
@@ -102,8 +102,9 @@ class Game(Base):
     """Модель игровой сессии (одиночный режим)."""
     __tablename__ = "games"
 
-    # ИСПРАВЛЕНО: autoincrement=True для SQLite совместимости
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    # ИСПРАВЛЕНО: Integer + autoincrement=True для SQLite совместимости
+    # Важно: используем Integer вместо BigInteger для корректного autoincrement в SQLite
+    id = Column(Integer, primary_key=True, autoincrement=True)
     user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False)
 
     # Настройки игры
